requirements-web.txt
flickrapi>=2.4.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.8.0
